Change product price from BigDecimal to double to make code simple

+++ src/main/java/ReceiptPrinter.java
@@ -1,4 +1,3 @@
-import java.math.BigInteger;
 
 public class ReceiptPrinter {
 
@@ -7,7 +6,7 @@ public class ReceiptPrinter {
 	}
 
 	public String oneItemInReceiptItem(Product product, int number) {
-		String str = String.format("名称：可口可乐，数量：%d瓶，单价：%1.2f（元），小计：%1.2f（元）",number,product.getPrice().doubleValue(),product.getPrice().doubleValue() * number);
+		String str = String.format("名称：可口可乐，数量：%d瓶，单价：%1.2f（元），小计：%1.2f（元）",number,product.getPrice(),product.getPrice() * number);
 		return str;
 	}
 }
